:recycle: move `copy_doc` to private
