+++ discord/commands/context.py
@@ -54,6 +54,7 @@ if TYPE_CHECKING:
     from typing import Callable, Awaitable
 
 from ..utils import cached_property
+from ..utils.private import copy_doc
 
 T = TypeVar("T")
 CogT = TypeVar("CogT", bound="Cog")
@@ -258,17 +259,17 @@ class ApplicationContext(discord.abc.Messageable):
         return None
 
     @property
-    @discord.utils.copy_doc(InteractionResponse.send_modal)
+    @copy_doc.copy_doc(InteractionResponse.send_modal)
     def send_modal(self) -> Callable[..., Awaitable[Interaction]]:
         return self.interaction.response.send_modal
 
     @property
-    @discord.utils.copy_doc(Interaction.respond)
+    @copy_doc.copy_doc(Interaction.respond)
     def respond(self, *args, **kwargs) -> Callable[..., Awaitable[Interaction | WebhookMessage]]:
         return self.interaction.respond
 
     @property
-    @discord.utils.copy_doc(InteractionResponse.send_message)
+    @copy_doc.copy_doc(InteractionResponse.send_message)
     def send_response(self) -> Callable[..., Awaitable[Interaction]]:
         if not self.interaction.response.is_done():
             return self.interaction.response.send_message
@@ -278,7 +279,7 @@ class ApplicationContext(discord.abc.Messageable):
             )
 
     @property
-    @discord.utils.copy_doc(Webhook.send)
+    @copy_doc.copy_doc(Webhook.send)
     def send_followup(self) -> Callable[..., Awaitable[WebhookMessage]]:
         if self.interaction.response.is_done():
             return self.followup.send
@@ -288,7 +289,7 @@ class ApplicationContext(discord.abc.Messageable):
             )
 
     @property
-    @discord.utils.copy_doc(InteractionResponse.defer)
+    @copy_doc.copy_doc(InteractionResponse.defer)
     def defer(self) -> Callable[..., Awaitable[None]]:
         return self.interaction.response.defer
 
@@ -322,7 +323,7 @@ class ApplicationContext(discord.abc.Messageable):
         return await self.interaction.delete_original_response(delay=delay)
 
     @property
-    @discord.utils.copy_doc(Interaction.edit_original_response)
+    @copy_doc.copy_doc(Interaction.edit_original_response)
     def edit(self) -> Callable[..., Awaitable[InteractionMessage]]:
         return self.interaction.edit_original_response
