+++ discord/ext/commands/bot.py
@@ -133,7 +133,7 @@ class BotBase(GroupMixin, discord.cog.CogMixin):
         self.help_command = DefaultHelpCommand() if help_command is MISSING else help_command
         self.strip_after_prefix = options.get("strip_after_prefix", False)
 
-    @discord.utils.copy_doc(discord.Client.close)
+    @discord.utils.private.copy_doc(discord.Client.close)
     async def close(self) -> None:
         for extension in tuple(self.__extensions):
             try:
